菜单权限控制

+++ server/service/userService.js
@@ -3,6 +3,7 @@ const md5 = require("blueimp-md5");
 const tableName = "sys_user";
 exports.login = (req, res) => {
     const { username, password } = req.body;
+    // SELECT u.user_id,u.username,u.`status`,r.menus from sys_user_role ur LEFT JOIN sys_user u ON u.user_id = ur.user_id LEFT JOIN sys_role r ON ur.role_id = r.role_id WHERE u.user_id = "125794949"
     base(tableName)
         .where({ username, password: md5(password) })
         .find()
@@ -11,14 +12,32 @@ exports.login = (req, res) => {
             if (JSON.stringify(user) !== "{}") {
                 // 生成一个cookie(userid: user._id), 并交给浏览器保存
                 res.cookie("userid", user.user_id, { maxAge: 1000 * 60 * 60 * 24 });
-                if (user.role_id) {
-                    // base('sys_user_role').join({table:""})
-                    res.send({ status: 0, data: user });
-                } else {
+                //sys_role  role_id menus
+                //sys_user_role user_id role_id
+                //sys_user user_id
+                if (user.username === "admin") {
                     delete user.password;
+                    user.menus = `["/home","/article","/category","/articles","/user","/role"]`;
+                    res.send({ status: 0, data: user });
+                    return new Promise(() => {});
+                } else {
+                    return base(
+                        `SELECT u.user_id,u.username,u.status,r.menus,u.email,u.mobile,u.create_user_id from sys_user_role ur 
+                LEFT JOIN sys_user u ON u.user_id = ur.user_id 
+                LEFT JOIN sys_role r ON ur.role_id = r.role_id 
+                WHERE u.user_id = ${user.user_id}`,
+                        false
+                    );
                     // 返回登陆成功信息(包含user)
-                    res.json({ status: 0, data: user });
                 }
+            } else {
+                res.json({ status: 1, msg: "该用户不存在" });
+                return new Promise(() => {});
+            }
+        })
+        .then((user) => {
+            if (user.length > 0) {
+                res.json({ status: 0, data: user[0] });
             } else {
                 res.json({ status: 1, msg: "该用户不存在" });
             }
@@ -55,19 +74,14 @@ exports.add = (req, res) => {
                 });
             } else {
                 res.json({ status: 1, msg: "此用户已存在" });
-                return new Promise(() => { });
+                return new Promise(() => {});
             }
         })
-        .then((user) => {
-            console.log(tableName, user);
-            base("sys_user_role")
-                .add({
-                    user_id: user,
-                    role_id: role,
-                })
-                .then((res) => {
-                    console.log("sys_user_role", res);
-                });
+        .then(async(user) => {
+            await base("sys_user_role").add({
+                user_id: user,
+                role_id: role,
+            });
             // 返回包含user的json数据
             res.json({ status: 0, data: user });
         })
@@ -98,16 +112,6 @@ exports.list = (req, res) => {
     base(tableName)
         .page(Number(pageNumber), Number(pageSize))
         .where(data)
-        // .field([
-        //     "user_id",
-        //     "username",
-        //     "status",
-        //     "email",
-        //     "mobile",
-        //     "create_user_id",
-        //     "create_time",
-        // ])
-        // .select()
         .countSelect()
         .then((list) => {
             res.json({
@@ -127,17 +131,12 @@ exports.delete = (req, res) => {
     base(tableName)
         .where({ user_id })
         .delete()
-        .then((response) => {
-            base("sys_user_role")
-                .where({ user_id })
-                .delete()
-                .then((res) => {
-                    console.log("sys_user_role", res);
-                });
+        .then(async(response) => {
+            await base("sys_user_role").where({ user_id }).delete();
             res.json({ status: 0, data: response });
         })
         .catch((error) => {
             console.error("删除异常", error);
             res.json({ status: 1, msg: "删除用户异常, 请重新尝试" });
         });
-}
+};
